Refactor device views into directory
