+++ src/v2/view-builder/views/device/DeviceChallengePollView.js
@@ -1,20 +1,20 @@
 /* global Promise */
 import { $, loc, createButton, View } from 'okta';
 import hbs from 'handlebars-inline-precompile';
-import BaseView from '../internals/BaseView';
-import BaseForm from '../internals/BaseForm';
-import BaseHeader from '../internals/BaseHeader';
-import HeaderBeacon from '../components/HeaderBeacon';
-import BaseFooter from '../internals//BaseFooter';
-import Logger from '../../../util/Logger';
-import DeviceFingerprint from '../../../util/DeviceFingerprint';
-import Util from '../../../util/Util';
-import Enums from '../../../util/Enums';
-import { CANCEL_POLLING_ACTION } from '../utils/Constants';
-import Link from '../components/Link';
-import { getIconClassNameForBeacon } from '../utils/AuthenticatorUtil';
-import { AUTHENTICATOR_KEY } from '../../ion/RemediationConstants';
-import BaseFormWithPolling from '../internals/BaseFormWithPolling';
+import BaseView from '../../internals/BaseView';
+import BaseForm from '../../internals/BaseForm';
+import BaseHeader from '../../internals/BaseHeader';
+import HeaderBeacon from '../../components/HeaderBeacon';
+import BaseFooter from '../../internals/BaseFooter';
+import Logger from '../../../../util/Logger';
+import DeviceFingerprint from '../../../../util/DeviceFingerprint';
+import Util from '../../../../util/Util';
+import Enums from '../../../../util/Enums';
+import { CANCEL_POLLING_ACTION } from '../../utils/Constants';
+import Link from '../../components/Link';
+import { getIconClassNameForBeacon } from '../../utils/AuthenticatorUtil';
+import { AUTHENTICATOR_KEY } from '../../../ion/RemediationConstants';
+import BaseFormWithPolling from '../../internals/BaseFormWithPolling';
 
 const request = (opts) => {
   const ajaxOptions = Object.assign({
